pyproject.toml
[project]
name = "scorix-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "accelerate>=1.11.0",
    "docx>=0.2.4",
    "dotenv>=0.9.9",
    "duckduckgo-search>=8.1.1",
    "elasticsearch>=9.1.1",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.128.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "google>=3.0.0",
    "google-generativeai>=0.8.5",
    "google-search-results>=2.4.2",
    "langchain>=1.0.2",
    "langchain-text-splitters>=1.0.0",
    "numpy>=2.3.4",
    "open-clip-torch>=3.2.0",
    "openai>=2.6.1",
    "openpyxl>=3.1.5",
    "pdfplumber>=0.11.7",
    "pillow>=12.0.0",
    "pymupdf>=1.26.5",
    "pypdf>=6.1.3",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
    "qdrant-client>=1.15.1",
    "redis>=7.0.0",
    "sentence-transformers>=5.1.1",
    "serpapi>=0.1.5",
    "streamlit>=1.52.2",
    "torch>=2.9.0",
    "transformers>=4.57.1",
    "wikipedia-api>=0.8.1",
]
